chore: renamed subpackages "acp_client" -> "sync_client" and "acp_async_client" -> "async_client".
