+++ acp_sdk/__init__.py
@@ -5,12 +5,12 @@ from typing import Literal
 type ACPSpecVersion = Literal["0.1"]
 CurrentACPSpecVersion = Literal["0.1"]
 
-from .acp_v0.acp_client.api import AgentsApi, RunsApi, ThreadsApi
-from .acp_v0.acp_client import ApiClient
-from .acp_v0.acp_async_client.api import AgentsApi as AsyncAgentsApi
-from .acp_v0.acp_async_client.api import RunsApi as AsyncRunsApi
-from .acp_v0.acp_async_client.api import ThreadsApi as AsyncThreadsApi
-from .acp_v0.acp_async_client import ApiClient as AsyncApiClient
+from .acp_v0.sync_client.api import AgentsApi, RunsApi, ThreadsApi
+from .acp_v0.sync_client import ApiClient
+from .acp_v0.async_client.api import AgentsApi as AsyncAgentsApi
+from .acp_v0.async_client.api import RunsApi as AsyncRunsApi
+from .acp_v0.async_client.api import ThreadsApi as AsyncThreadsApi
+from .acp_v0.async_client import ApiClient as AsyncApiClient
 from .acp_v0 import ApiResponse, Configuration
 from .acp_v0.spec_version import ACP_VERSION, ACP_MAJOR_VERSION, ACP_MINOR_VERSION
