--- acp_sdk/acp_v0/acp_client/api/__init__.py
@@ -1,9 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025 Cisco and/or its affiliates.
-# SPDX-License-Identifier: Apache-2.0
-# flake8: noqa
-
-# import apis into api package
-from acp_sdk.acp_v0.acp_client.api.agents_api import AgentsApi
-from acp_sdk.acp_v0.acp_client.api.runs_api import RunsApi
-from acp_sdk.acp_v0.acp_client.api.threads_api import ThreadsApi
-
