+++ acp_sdk/acp_v0/async_client/api/runs_api.py
@@ -18,20 +18,21 @@ from pydantic import validate_call, Field, StrictFloat, StrictStr, StrictInt
 from typing import Any, Dict, List, Optional, Tuple, Union
 from typing_extensions import Annotated
 
-from pydantic import Field, StrictStr
-from typing import Any, Dict, List
+from pydantic import Field, StrictInt, StrictStr
+from typing import Any, Dict, List, Optional
 from typing_extensions import Annotated
 from acp_sdk.acp_v0.models.run import Run
-from acp_sdk.acp_v0.models.thread import Thread
-from acp_sdk.acp_v0.models.thread_create import ThreadCreate
-from acp_sdk.acp_v0.models.thread_search_request import ThreadSearchRequest
+from acp_sdk.acp_v0.models.run_create import RunCreate
+from acp_sdk.acp_v0.models.run_output import RunOutput
+from acp_sdk.acp_v0.models.run_output_stream import RunOutputStream
+from acp_sdk.acp_v0.models.run_search_request import RunSearchRequest
 
-from acp_sdk.acp_v0.acp_async_client.api_client import ApiClient, RequestSerialized
+from acp_sdk.acp_v0.async_client.api_client import ApiClient, RequestSerialized
 from acp_sdk.acp_v0.api_response import ApiResponse
-from acp_sdk.acp_v0.acp_async_client.rest import RESTResponseType
+from acp_sdk.acp_v0.async_client.rest import RESTResponseType
 
 
-class ThreadsApi:
+class RunsApi:
     """NOTE: This class is auto generated by OpenAPI Generator
     Ref: https://openapi-generator.tech
 
@@ -45,9 +46,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def create_thread(
+    async def create_run(
         self,
-        thread_create: ThreadCreate,
+        run_create: RunCreate,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -60,13 +61,13 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> Thread:
-        """Create an empty Thread
+    ) -> Run:
+        """Create Background Run
 
-        Create an empty thread. This is useful to associate metadata to a thread.
+        Create a run, return the run descriptor immediately. Don't wait for the final run output.
 
-        :param thread_create: (required)
-        :type thread_create: ThreadCreate
+        :param run_create: (required)
+        :type run_create: RunCreate
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -89,8 +90,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._create_thread_serialize(
-            thread_create=thread_create,
+        _param = self._create_run_serialize(
+            run_create=run_create,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -98,7 +99,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -115,9 +116,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def create_thread_with_http_info(
+    async def create_run_with_http_info(
         self,
-        thread_create: ThreadCreate,
+        run_create: RunCreate,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -130,13 +131,13 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[Thread]:
-        """Create an empty Thread
+    ) -> ApiResponse[Run]:
+        """Create Background Run
 
-        Create an empty thread. This is useful to associate metadata to a thread.
+        Create a run, return the run descriptor immediately. Don't wait for the final run output.
 
-        :param thread_create: (required)
-        :type thread_create: ThreadCreate
+        :param run_create: (required)
+        :type run_create: RunCreate
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -159,8 +160,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._create_thread_serialize(
-            thread_create=thread_create,
+        _param = self._create_run_serialize(
+            run_create=run_create,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -168,7 +169,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -185,9 +186,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def create_thread_without_preload_content(
+    async def create_run_without_preload_content(
         self,
-        thread_create: ThreadCreate,
+        run_create: RunCreate,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -201,12 +202,12 @@ class ThreadsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Create an empty Thread
+        """Create Background Run
 
-        Create an empty thread. This is useful to associate metadata to a thread.
+        Create a run, return the run descriptor immediately. Don't wait for the final run output.
 
-        :param thread_create: (required)
-        :type thread_create: ThreadCreate
+        :param run_create: (required)
+        :type run_create: RunCreate
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -229,8 +230,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._create_thread_serialize(
-            thread_create=thread_create,
+        _param = self._create_run_serialize(
+            run_create=run_create,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -238,7 +239,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -250,9 +251,9 @@ class ThreadsApi:
         return response_data.response
 
 
-    def _create_thread_serialize(
+    def _create_run_serialize(
         self,
-        thread_create,
+        run_create,
         _request_auth,
         _content_type,
         _headers,
@@ -278,8 +279,8 @@ class ThreadsApi:
         # process the header parameters
         # process the form parameters
         # process the body parameter
-        if thread_create is not None:
-            _body_params = thread_create
+        if run_create is not None:
+            _body_params = run_create
 
 
         # set the HTTP header `Accept`
@@ -310,7 +311,7 @@ class ThreadsApi:
 
         return self.api_client.param_serialize(
             method='POST',
-            resource_path='/threads',
+            resource_path='/runs',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -327,9 +328,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def delete_thread(
+    async def delete_run(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -342,13 +343,13 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> Thread:
-        """Delete a thread. If the thread contains any pending run, deletion fails.
+    ) -> Run:
+        """Delete a run. If running, cancel and then delete.
 
-        Delete a thread.
+        Cancel a run.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the agent. (required)
+        :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -371,8 +372,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._delete_thread_serialize(
-            thread_id=thread_id,
+        _param = self._delete_run_serialize(
+            run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -380,7 +381,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -397,9 +398,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def delete_thread_with_http_info(
+    async def delete_run_with_http_info(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -412,13 +413,13 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[Thread]:
-        """Delete a thread. If the thread contains any pending run, deletion fails.
+    ) -> ApiResponse[Run]:
+        """Delete a run. If running, cancel and then delete.
 
-        Delete a thread.
+        Cancel a run.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the agent. (required)
+        :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -441,8 +442,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._delete_thread_serialize(
-            thread_id=thread_id,
+        _param = self._delete_run_serialize(
+            run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -450,7 +451,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -467,9 +468,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def delete_thread_without_preload_content(
+    async def delete_run_without_preload_content(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -483,12 +484,12 @@ class ThreadsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Delete a thread. If the thread contains any pending run, deletion fails.
+        """Delete a run. If running, cancel and then delete.
 
-        Delete a thread.
+        Cancel a run.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the agent. (required)
+        :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -511,8 +512,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._delete_thread_serialize(
-            thread_id=thread_id,
+        _param = self._delete_run_serialize(
+            run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -520,7 +521,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -532,9 +533,9 @@ class ThreadsApi:
         return response_data.response
 
 
-    def _delete_thread_serialize(
+    def _delete_run_serialize(
         self,
-        thread_id,
+        run_id,
         _request_auth,
         _content_type,
         _headers,
@@ -556,8 +557,8 @@ class ThreadsApi:
         _body_params: Optional[bytes] = None
 
         # process the path parameters
-        if thread_id is not None:
-            _path_params['thread_id'] = thread_id
+        if run_id is not None:
+            _path_params['run_id'] = run_id
         # process the query parameters
         # process the header parameters
         # process the form parameters
@@ -579,7 +580,7 @@ class ThreadsApi:
 
         return self.api_client.param_serialize(
             method='DELETE',
-            resource_path='/threads/{thread_id}',
+            resource_path='/runs/{run_id}',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -596,9 +597,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_run_threadstate(
+    async def get_run(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -611,12 +612,12 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> object:
-        """Retrieve the thread state at the end of the run
+    ) -> Run:
+        """Get a previously created Run
 
-        This call can be used only for agents that support thread, i.e. for Runs that specify a thread ID. It can be called only on runs that are in `success` status. It returns the thread state at the end of the Run. Can be used to reconstruct the evolution of the thread state in its history.
+        Get a run from its ID. Don't wait for the final run output.
 
-        :param run_id: The ID of the run. (required)
+        :param run_id: The ID of the agent. (required)
         :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
@@ -640,7 +641,7 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_threadstate_serialize(
+        _param = self._get_run_serialize(
             run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
@@ -649,7 +650,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "object",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -666,9 +667,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_run_threadstate_with_http_info(
+    async def get_run_with_http_info(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -681,12 +682,12 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[object]:
-        """Retrieve the thread state at the end of the run
+    ) -> ApiResponse[Run]:
+        """Get a previously created Run
 
-        This call can be used only for agents that support thread, i.e. for Runs that specify a thread ID. It can be called only on runs that are in `success` status. It returns the thread state at the end of the Run. Can be used to reconstruct the evolution of the thread state in its history.
+        Get a run from its ID. Don't wait for the final run output.
 
-        :param run_id: The ID of the run. (required)
+        :param run_id: The ID of the agent. (required)
         :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
@@ -710,7 +711,7 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_threadstate_serialize(
+        _param = self._get_run_serialize(
             run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
@@ -719,7 +720,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "object",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -736,9 +737,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_run_threadstate_without_preload_content(
+    async def get_run_without_preload_content(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -752,11 +753,11 @@ class ThreadsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Retrieve the thread state at the end of the run
+        """Get a previously created Run
 
-        This call can be used only for agents that support thread, i.e. for Runs that specify a thread ID. It can be called only on runs that are in `success` status. It returns the thread state at the end of the Run. Can be used to reconstruct the evolution of the thread state in its history.
+        Get a run from its ID. Don't wait for the final run output.
 
-        :param run_id: The ID of the run. (required)
+        :param run_id: The ID of the agent. (required)
         :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
@@ -780,7 +781,7 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_threadstate_serialize(
+        _param = self._get_run_serialize(
             run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
@@ -789,7 +790,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "object",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -801,7 +802,7 @@ class ThreadsApi:
         return response_data.response
 
 
-    def _get_run_threadstate_serialize(
+    def _get_run_serialize(
         self,
         run_id,
         _request_auth,
@@ -848,7 +849,7 @@ class ThreadsApi:
 
         return self.api_client.param_serialize(
             method='GET',
-            resource_path='/runs/{run_id}/threadstate',
+            resource_path='/runs/{run_id}',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -865,9 +866,10 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread(
+    async def get_run_output(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
+        block_timeout: Optional[StrictInt] = None,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -880,13 +882,15 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> Thread:
-        """Get a previously created Thread
+    ) -> RunOutput:
+        """Retrieve last output of a run if available
 
-        Get a thread from its ID. 
+        Retrieve the last output of the run.  The output can be:   * an interrupt, this happens when the agent run status is `interrupted`   * the final result of the run, this happens when the agent run status is `success`   * an error, this happens when the agent run status is `error` or `timeout`   If the block timeout is provided and the current run status is `pending`, this call blocks until the state changes or the timeout expires.  If no timeout is provided or the timeout has expired and  run status is `pending`, this call returns `204` with no content.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the run. (required)
+        :type run_id: str
+        :param block_timeout:
+        :type block_timeout: int
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -909,8 +913,9 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_serialize(
-            thread_id=thread_id,
+        _param = self._get_run_output_serialize(
+            run_id=run_id,
+            block_timeout=block_timeout,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -918,7 +923,8 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "RunOutput",
+            '204': None,
             '404': "str",
             '409': "str",
             '422': "str",
@@ -935,9 +941,10 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_with_http_info(
+    async def get_run_output_with_http_info(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
+        block_timeout: Optional[StrictInt] = None,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -950,13 +957,15 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[Thread]:
-        """Get a previously created Thread
+    ) -> ApiResponse[RunOutput]:
+        """Retrieve last output of a run if available
 
-        Get a thread from its ID. 
+        Retrieve the last output of the run.  The output can be:   * an interrupt, this happens when the agent run status is `interrupted`   * the final result of the run, this happens when the agent run status is `success`   * an error, this happens when the agent run status is `error` or `timeout`   If the block timeout is provided and the current run status is `pending`, this call blocks until the state changes or the timeout expires.  If no timeout is provided or the timeout has expired and  run status is `pending`, this call returns `204` with no content.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the run. (required)
+        :type run_id: str
+        :param block_timeout:
+        :type block_timeout: int
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -979,8 +988,9 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_serialize(
-            thread_id=thread_id,
+        _param = self._get_run_output_serialize(
+            run_id=run_id,
+            block_timeout=block_timeout,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -988,7 +998,8 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "RunOutput",
+            '204': None,
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1005,9 +1016,10 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_without_preload_content(
+    async def get_run_output_without_preload_content(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
+        block_timeout: Optional[StrictInt] = None,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1021,12 +1033,14 @@ class ThreadsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Get a previously created Thread
+        """Retrieve last output of a run if available
 
-        Get a thread from its ID. 
+        Retrieve the last output of the run.  The output can be:   * an interrupt, this happens when the agent run status is `interrupted`   * the final result of the run, this happens when the agent run status is `success`   * an error, this happens when the agent run status is `error` or `timeout`   If the block timeout is provided and the current run status is `pending`, this call blocks until the state changes or the timeout expires.  If no timeout is provided or the timeout has expired and  run status is `pending`, this call returns `204` with no content.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the run. (required)
+        :type run_id: str
+        :param block_timeout:
+        :type block_timeout: int
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1049,8 +1063,9 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_serialize(
-            thread_id=thread_id,
+        _param = self._get_run_output_serialize(
+            run_id=run_id,
+            block_timeout=block_timeout,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1058,7 +1073,8 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Thread",
+            '200': "RunOutput",
+            '204': None,
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1070,9 +1086,10 @@ class ThreadsApi:
         return response_data.response
 
 
-    def _get_thread_serialize(
+    def _get_run_output_serialize(
         self,
-        thread_id,
+        run_id,
+        block_timeout,
         _request_auth,
         _content_type,
         _headers,
@@ -1094,9 +1111,13 @@ class ThreadsApi:
         _body_params: Optional[bytes] = None
 
         # process the path parameters
-        if thread_id is not None:
-            _path_params['thread_id'] = thread_id
+        if run_id is not None:
+            _path_params['run_id'] = run_id
         # process the query parameters
+        if block_timeout is not None:
+            
+            _query_params.append(('block_timeout', block_timeout))
+            
         # process the header parameters
         # process the form parameters
         # process the body parameter
@@ -1117,7 +1138,7 @@ class ThreadsApi:
 
         return self.api_client.param_serialize(
             method='GET',
-            resource_path='/threads/{thread_id}',
+            resource_path='/runs/{run_id}/output',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -1134,9 +1155,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_history(
+    async def get_run_stream(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1149,13 +1170,13 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> List[Run]:
-        """Retrieve the list of runs and associated state at the end of each run.
+    ) -> RunOutputStream:
+        """Stream the run output
 
-        Retrieve ordered list of runs for this thread in chronological order.
+        Send a stream of events using Server-sent events (SEE). See <https://html.spec.whatwg.org/multipage/server-sent-events.html> for details.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the run. (required)
+        :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1178,8 +1199,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_history_serialize(
-            thread_id=thread_id,
+        _param = self._get_run_stream_serialize(
+            run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1187,7 +1208,8 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "List[Run]",
+            '200': "RunOutputStream",
+            '204': None,
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1204,9 +1226,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_history_with_http_info(
+    async def get_run_stream_with_http_info(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1219,13 +1241,13 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[List[Run]]:
-        """Retrieve the list of runs and associated state at the end of each run.
+    ) -> ApiResponse[RunOutputStream]:
+        """Stream the run output
 
-        Retrieve ordered list of runs for this thread in chronological order.
+        Send a stream of events using Server-sent events (SEE). See <https://html.spec.whatwg.org/multipage/server-sent-events.html> for details.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the run. (required)
+        :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1248,8 +1270,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_history_serialize(
-            thread_id=thread_id,
+        _param = self._get_run_stream_serialize(
+            run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1257,7 +1279,8 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "List[Run]",
+            '200': "RunOutputStream",
+            '204': None,
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1274,9 +1297,9 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_history_without_preload_content(
+    async def get_run_stream_without_preload_content(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1290,12 +1313,12 @@ class ThreadsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Retrieve the list of runs and associated state at the end of each run.
+        """Stream the run output
 
-        Retrieve ordered list of runs for this thread in chronological order.
+        Send a stream of events using Server-sent events (SEE). See <https://html.spec.whatwg.org/multipage/server-sent-events.html> for details.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the run. (required)
+        :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1318,8 +1341,8 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_history_serialize(
-            thread_id=thread_id,
+        _param = self._get_run_stream_serialize(
+            run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1327,7 +1350,8 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "List[Run]",
+            '200': "RunOutputStream",
+            '204': None,
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1339,9 +1363,9 @@ class ThreadsApi:
         return response_data.response
 
 
-    def _get_thread_history_serialize(
+    def _get_run_stream_serialize(
         self,
-        thread_id,
+        run_id,
         _request_auth,
         _content_type,
         _headers,
@@ -1363,8 +1387,8 @@ class ThreadsApi:
         _body_params: Optional[bytes] = None
 
         # process the path parameters
-        if thread_id is not None:
-            _path_params['thread_id'] = thread_id
+        if run_id is not None:
+            _path_params['run_id'] = run_id
         # process the query parameters
         # process the header parameters
         # process the form parameters
@@ -1375,6 +1399,7 @@ class ThreadsApi:
         if 'Accept' not in _header_params:
             _header_params['Accept'] = self.api_client.select_header_accept(
                 [
+                    'text/event-stream', 
                     'application/json'
                 ]
             )
@@ -1386,7 +1411,7 @@ class ThreadsApi:
 
         return self.api_client.param_serialize(
             method='GET',
-            resource_path='/threads/{thread_id}/history',
+            resource_path='/runs/{run_id}/stream',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -1403,9 +1428,10 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_state(
+    async def resume_run(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        body: Dict[str, Any],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1418,13 +1444,15 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> object:
-        """Retrieve the current state associated with the thread
+    ) -> Run:
+        """Resume an interrupted Run
 
-        Retrieve the the current state associated with the thread
+        Provide the needed input to a run to resume its execution. Can only be called for runs that are in the interrupted state Schema of the provided input must match with the schema specified in the agent specs under interrupts for the interrupt type the agent generated for this specific interruption.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the agent. (required)
+        :type run_id: str
+        :param body: (required)
+        :type body: object
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1447,8 +1475,9 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_state_serialize(
-            thread_id=thread_id,
+        _param = self._resume_run_serialize(
+            run_id=run_id,
+            body=body,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1456,7 +1485,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "object",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1473,9 +1502,10 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_state_with_http_info(
+    async def resume_run_with_http_info(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        body: Dict[str, Any],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1488,13 +1518,15 @@ class ThreadsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[object]:
-        """Retrieve the current state associated with the thread
+    ) -> ApiResponse[Run]:
+        """Resume an interrupted Run
 
-        Retrieve the the current state associated with the thread
+        Provide the needed input to a run to resume its execution. Can only be called for runs that are in the interrupted state Schema of the provided input must match with the schema specified in the agent specs under interrupts for the interrupt type the agent generated for this specific interruption.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the agent. (required)
+        :type run_id: str
+        :param body: (required)
+        :type body: object
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1517,8 +1549,9 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_state_serialize(
-            thread_id=thread_id,
+        _param = self._resume_run_serialize(
+            run_id=run_id,
+            body=body,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1526,7 +1559,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "object",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1543,9 +1576,10 @@ class ThreadsApi:
 
 
     @validate_call
-    async def get_thread_state_without_preload_content(
+    async def resume_run_without_preload_content(
         self,
-        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        body: Dict[str, Any],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1559,12 +1593,14 @@ class ThreadsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Retrieve the current state associated with the thread
+        """Resume an interrupted Run
 
-        Retrieve the the current state associated with the thread
+        Provide the needed input to a run to resume its execution. Can only be called for runs that are in the interrupted state Schema of the provided input must match with the schema specified in the agent specs under interrupts for the interrupt type the agent generated for this specific interruption.
 
-        :param thread_id: The ID of the thread. (required)
-        :type thread_id: str
+        :param run_id: The ID of the agent. (required)
+        :type run_id: str
+        :param body: (required)
+        :type body: object
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1587,8 +1623,9 @@ class ThreadsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_thread_state_serialize(
-            thread_id=thread_id,
+        _param = self._resume_run_serialize(
+            run_id=run_id,
+            body=body,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1596,7 +1633,7 @@ class ThreadsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "object",
+            '200': "Run",
             '404': "str",
             '409': "str",
             '422': "str",
@@ -1608,9 +1645,10 @@ class ThreadsApi:
         return response_data.response
 
 
-    def _get_thread_state_serialize(
+    def _resume_run_serialize(
         self,
-        thread_id,
+        run_id,
+        body,
         _request_auth,
         _content_type,
         _headers,
@@ -1632,275 +1670,14 @@ class ThreadsApi:
         _body_params: Optional[bytes] = None
 
         # process the path parameters
-        if thread_id is not None:
-            _path_params['thread_id'] = thread_id
+        if run_id is not None:
+            _path_params['run_id'] = run_id
         # process the query parameters
         # process the header parameters
         # process the form parameters
         # process the body parameter
-
-
-        # set the HTTP header `Accept`
-        if 'Accept' not in _header_params:
-            _header_params['Accept'] = self.api_client.select_header_accept(
-                [
-                    'application/json'
-                ]
-            )
-
-
-        # authentication setting
-        _auth_settings: List[str] = [
-        ]
-
-        return self.api_client.param_serialize(
-            method='GET',
-            resource_path='/threads/{thread_id}/state',
-            path_params=_path_params,
-            query_params=_query_params,
-            header_params=_header_params,
-            body=_body_params,
-            post_params=_form_params,
-            files=_files,
-            auth_settings=_auth_settings,
-            collection_formats=_collection_formats,
-            _host=_host,
-            _request_auth=_request_auth
-        )
-
-
-
-
-    @validate_call
-    async def search_threads(
-        self,
-        thread_search_request: ThreadSearchRequest,
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> List[Thread]:
-        """Search Threads
-
-        Search for threads.  This endpoint also functions as the endpoint to list all threads.
-
-        :param thread_search_request: (required)
-        :type thread_search_request: ThreadSearchRequest
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._search_threads_serialize(
-            thread_search_request=thread_search_request,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "List[Thread]",
-            '422': "str",
-        }
-        response_data = await self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        await response_data.read()
-        return self.api_client.response_deserialize(
-            response_data=response_data,
-            response_types_map=_response_types_map,
-        ).data
-
-
-    @validate_call
-    async def search_threads_with_http_info(
-        self,
-        thread_search_request: ThreadSearchRequest,
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[List[Thread]]:
-        """Search Threads
-
-        Search for threads.  This endpoint also functions as the endpoint to list all threads.
-
-        :param thread_search_request: (required)
-        :type thread_search_request: ThreadSearchRequest
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._search_threads_serialize(
-            thread_search_request=thread_search_request,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "List[Thread]",
-            '422': "str",
-        }
-        response_data = await self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        await response_data.read()
-        return self.api_client.response_deserialize(
-            response_data=response_data,
-            response_types_map=_response_types_map,
-        )
-
-
-    @validate_call
-    async def search_threads_without_preload_content(
-        self,
-        thread_search_request: ThreadSearchRequest,
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> RESTResponseType:
-        """Search Threads
-
-        Search for threads.  This endpoint also functions as the endpoint to list all threads.
-
-        :param thread_search_request: (required)
-        :type thread_search_request: ThreadSearchRequest
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._search_threads_serialize(
-            thread_search_request=thread_search_request,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "List[Thread]",
-            '422': "str",
-        }
-        response_data = await self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        return response_data.response
-
-
-    def _search_threads_serialize(
-        self,
-        thread_search_request,
-        _request_auth,
-        _content_type,
-        _headers,
-        _host_index,
-    ) -> RequestSerialized:
-
-        _host = None
-
-        _collection_formats: Dict[str, str] = {
-        }
-
-        _path_params: Dict[str, str] = {}
-        _query_params: List[Tuple[str, str]] = []
-        _header_params: Dict[str, Optional[str]] = _headers or {}
-        _form_params: List[Tuple[str, str]] = []
-        _files: Dict[
-            str, Union[str, bytes, List[str], List[bytes], List[Tuple[str, bytes]]]
-        ] = {}
-        _body_params: Optional[bytes] = None
-
-        # process the path parameters
-        # process the query parameters
-        # process the header parameters
-        # process the form parameters
-        # process the body parameter
-        if thread_search_request is not None:
-            _body_params = thread_search_request
+        if body is not None:
+            _body_params = body
 
 
         # set the HTTP header `Accept`
@@ -1931,7 +1708,283 @@ class ThreadsApi:
 
         return self.api_client.param_serialize(
             method='POST',
-            resource_path='/threads/search',
+            resource_path='/runs/{run_id}',
+            path_params=_path_params,
+            query_params=_query_params,
+            header_params=_header_params,
+            body=_body_params,
+            post_params=_form_params,
+            files=_files,
+            auth_settings=_auth_settings,
+            collection_formats=_collection_formats,
+            _host=_host,
+            _request_auth=_request_auth
+        )
+
+
+
+
+    @validate_call
+    async def search_runs(
+        self,
+        run_search_request: RunSearchRequest,
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> List[Run]:
+        """Search Runs
+
+        Search for runs.  This endpoint also functions as the endpoint to list all runs.
+
+        :param run_search_request: (required)
+        :type run_search_request: RunSearchRequest
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._search_runs_serialize(
+            run_search_request=run_search_request,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "List[Run]",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        await response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        ).data
+
+
+    @validate_call
+    async def search_runs_with_http_info(
+        self,
+        run_search_request: RunSearchRequest,
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> ApiResponse[List[Run]]:
+        """Search Runs
+
+        Search for runs.  This endpoint also functions as the endpoint to list all runs.
+
+        :param run_search_request: (required)
+        :type run_search_request: RunSearchRequest
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._search_runs_serialize(
+            run_search_request=run_search_request,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "List[Run]",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        await response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        )
+
+
+    @validate_call
+    async def search_runs_without_preload_content(
+        self,
+        run_search_request: RunSearchRequest,
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> RESTResponseType:
+        """Search Runs
+
+        Search for runs.  This endpoint also functions as the endpoint to list all runs.
+
+        :param run_search_request: (required)
+        :type run_search_request: RunSearchRequest
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._search_runs_serialize(
+            run_search_request=run_search_request,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "List[Run]",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        return response_data.response
+
+
+    def _search_runs_serialize(
+        self,
+        run_search_request,
+        _request_auth,
+        _content_type,
+        _headers,
+        _host_index,
+    ) -> RequestSerialized:
+
+        _host = None
+
+        _collection_formats: Dict[str, str] = {
+        }
+
+        _path_params: Dict[str, str] = {}
+        _query_params: List[Tuple[str, str]] = []
+        _header_params: Dict[str, Optional[str]] = _headers or {}
+        _form_params: List[Tuple[str, str]] = []
+        _files: Dict[
+            str, Union[str, bytes, List[str], List[bytes], List[Tuple[str, bytes]]]
+        ] = {}
+        _body_params: Optional[bytes] = None
+
+        # process the path parameters
+        # process the query parameters
+        # process the header parameters
+        # process the form parameters
+        # process the body parameter
+        if run_search_request is not None:
+            _body_params = run_search_request
+
+
+        # set the HTTP header `Accept`
+        if 'Accept' not in _header_params:
+            _header_params['Accept'] = self.api_client.select_header_accept(
+                [
+                    'application/json'
+                ]
+            )
+
+        # set the HTTP header `Content-Type`
+        if _content_type:
+            _header_params['Content-Type'] = _content_type
+        else:
+            _default_content_type = (
+                self.api_client.select_header_content_type(
+                    [
+                        'application/json'
+                    ]
+                )
+            )
+            if _default_content_type is not None:
+                _header_params['Content-Type'] = _default_content_type
+
+        # authentication setting
+        _auth_settings: List[str] = [
+        ]
+
+        return self.api_client.param_serialize(
+            method='POST',
+            resource_path='/runs/search',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
