+++ acp_sdk/acp_v0/async_client/api/threads_api.py
@@ -18,21 +18,20 @@ from pydantic import validate_call, Field, StrictFloat, StrictStr, StrictInt
 from typing import Any, Dict, List, Optional, Tuple, Union
 from typing_extensions import Annotated
 
-from pydantic import Field, StrictInt, StrictStr
-from typing import Any, Dict, List, Optional
+from pydantic import Field, StrictStr
+from typing import Any, Dict, List
 from typing_extensions import Annotated
 from acp_sdk.acp_v0.models.run import Run
-from acp_sdk.acp_v0.models.run_create import RunCreate
-from acp_sdk.acp_v0.models.run_output import RunOutput
-from acp_sdk.acp_v0.models.run_output_stream import RunOutputStream
-from acp_sdk.acp_v0.models.run_search_request import RunSearchRequest
+from acp_sdk.acp_v0.models.thread import Thread
+from acp_sdk.acp_v0.models.thread_create import ThreadCreate
+from acp_sdk.acp_v0.models.thread_search_request import ThreadSearchRequest
 
-from acp_sdk.acp_v0.acp_client.api_client import ApiClient, RequestSerialized
+from acp_sdk.acp_v0.async_client.api_client import ApiClient, RequestSerialized
 from acp_sdk.acp_v0.api_response import ApiResponse
-from acp_sdk.acp_v0.acp_client.rest import RESTResponseType
+from acp_sdk.acp_v0.async_client.rest import RESTResponseType
 
 
-class RunsApi:
+class ThreadsApi:
     """NOTE: This class is auto generated by OpenAPI Generator
     Ref: https://openapi-generator.tech
 
@@ -46,9 +45,9 @@ class RunsApi:
 
 
     @validate_call
-    def create_run(
+    async def create_thread(
         self,
-        run_create: RunCreate,
+        thread_create: ThreadCreate,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -61,13 +60,13 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> Run:
-        """Create Background Run
+    ) -> Thread:
+        """Create an empty Thread
 
-        Create a run, return the run descriptor immediately. Don't wait for the final run output.
+        Create an empty thread. This is useful to associate metadata to a thread.
 
-        :param run_create: (required)
-        :type run_create: RunCreate
+        :param thread_create: (required)
+        :type thread_create: ThreadCreate
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -90,8 +89,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._create_run_serialize(
-            run_create=run_create,
+        _param = self._create_thread_serialize(
+            thread_create=thread_create,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -99,16 +98,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -116,9 +115,9 @@ class RunsApi:
 
 
     @validate_call
-    def create_run_with_http_info(
+    async def create_thread_with_http_info(
         self,
-        run_create: RunCreate,
+        thread_create: ThreadCreate,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -131,13 +130,13 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[Run]:
-        """Create Background Run
+    ) -> ApiResponse[Thread]:
+        """Create an empty Thread
 
-        Create a run, return the run descriptor immediately. Don't wait for the final run output.
+        Create an empty thread. This is useful to associate metadata to a thread.
 
-        :param run_create: (required)
-        :type run_create: RunCreate
+        :param thread_create: (required)
+        :type thread_create: ThreadCreate
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -160,8 +159,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._create_run_serialize(
-            run_create=run_create,
+        _param = self._create_thread_serialize(
+            thread_create=thread_create,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -169,16 +168,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -186,9 +185,9 @@ class RunsApi:
 
 
     @validate_call
-    def create_run_without_preload_content(
+    async def create_thread_without_preload_content(
         self,
-        run_create: RunCreate,
+        thread_create: ThreadCreate,
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -202,12 +201,12 @@ class RunsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Create Background Run
+        """Create an empty Thread
 
-        Create a run, return the run descriptor immediately. Don't wait for the final run output.
+        Create an empty thread. This is useful to associate metadata to a thread.
 
-        :param run_create: (required)
-        :type run_create: RunCreate
+        :param thread_create: (required)
+        :type thread_create: ThreadCreate
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -230,8 +229,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._create_run_serialize(
-            run_create=run_create,
+        _param = self._create_thread_serialize(
+            thread_create=thread_create,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -239,21 +238,21 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
         return response_data.response
 
 
-    def _create_run_serialize(
+    def _create_thread_serialize(
         self,
-        run_create,
+        thread_create,
         _request_auth,
         _content_type,
         _headers,
@@ -279,8 +278,8 @@ class RunsApi:
         # process the header parameters
         # process the form parameters
         # process the body parameter
-        if run_create is not None:
-            _body_params = run_create
+        if thread_create is not None:
+            _body_params = thread_create
 
 
         # set the HTTP header `Accept`
@@ -311,7 +310,7 @@ class RunsApi:
 
         return self.api_client.param_serialize(
             method='POST',
-            resource_path='/runs',
+            resource_path='/threads',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -328,9 +327,9 @@ class RunsApi:
 
 
     @validate_call
-    def delete_run(
+    async def delete_thread(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -343,13 +342,13 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> Run:
-        """Delete a run. If running, cancel and then delete.
+    ) -> Thread:
+        """Delete a thread. If the thread contains any pending run, deletion fails.
 
-        Cancel a run.
+        Delete a thread.
 
-        :param run_id: The ID of the agent. (required)
-        :type run_id: str
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -372,8 +371,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._delete_run_serialize(
-            run_id=run_id,
+        _param = self._delete_thread_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -381,16 +380,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -398,9 +397,9 @@ class RunsApi:
 
 
     @validate_call
-    def delete_run_with_http_info(
+    async def delete_thread_with_http_info(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -413,13 +412,13 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[Run]:
-        """Delete a run. If running, cancel and then delete.
+    ) -> ApiResponse[Thread]:
+        """Delete a thread. If the thread contains any pending run, deletion fails.
 
-        Cancel a run.
+        Delete a thread.
 
-        :param run_id: The ID of the agent. (required)
-        :type run_id: str
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -442,8 +441,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._delete_run_serialize(
-            run_id=run_id,
+        _param = self._delete_thread_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -451,16 +450,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -468,9 +467,9 @@ class RunsApi:
 
 
     @validate_call
-    def delete_run_without_preload_content(
+    async def delete_thread_without_preload_content(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -484,12 +483,12 @@ class RunsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Delete a run. If running, cancel and then delete.
+        """Delete a thread. If the thread contains any pending run, deletion fails.
 
-        Cancel a run.
+        Delete a thread.
 
-        :param run_id: The ID of the agent. (required)
-        :type run_id: str
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -512,8 +511,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._delete_run_serialize(
-            run_id=run_id,
+        _param = self._delete_thread_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -521,21 +520,21 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
         return response_data.response
 
 
-    def _delete_run_serialize(
+    def _delete_thread_serialize(
         self,
-        run_id,
+        thread_id,
         _request_auth,
         _content_type,
         _headers,
@@ -557,8 +556,8 @@ class RunsApi:
         _body_params: Optional[bytes] = None
 
         # process the path parameters
-        if run_id is not None:
-            _path_params['run_id'] = run_id
+        if thread_id is not None:
+            _path_params['thread_id'] = thread_id
         # process the query parameters
         # process the header parameters
         # process the form parameters
@@ -580,7 +579,7 @@ class RunsApi:
 
         return self.api_client.param_serialize(
             method='DELETE',
-            resource_path='/runs/{run_id}',
+            resource_path='/threads/{thread_id}',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -597,9 +596,9 @@ class RunsApi:
 
 
     @validate_call
-    def get_run(
+    async def get_run_threadstate(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -612,12 +611,12 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> Run:
-        """Get a previously created Run
+    ) -> object:
+        """Retrieve the thread state at the end of the run
 
-        Get a run from its ID. Don't wait for the final run output.
+        This call can be used only for agents that support thread, i.e. for Runs that specify a thread ID. It can be called only on runs that are in `success` status. It returns the thread state at the end of the Run. Can be used to reconstruct the evolution of the thread state in its history.
 
-        :param run_id: The ID of the agent. (required)
+        :param run_id: The ID of the run. (required)
         :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
@@ -641,7 +640,7 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_serialize(
+        _param = self._get_run_threadstate_serialize(
             run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
@@ -650,16 +649,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "object",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -667,9 +666,9 @@ class RunsApi:
 
 
     @validate_call
-    def get_run_with_http_info(
+    async def get_run_threadstate_with_http_info(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -682,12 +681,12 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[Run]:
-        """Get a previously created Run
+    ) -> ApiResponse[object]:
+        """Retrieve the thread state at the end of the run
 
-        Get a run from its ID. Don't wait for the final run output.
+        This call can be used only for agents that support thread, i.e. for Runs that specify a thread ID. It can be called only on runs that are in `success` status. It returns the thread state at the end of the Run. Can be used to reconstruct the evolution of the thread state in its history.
 
-        :param run_id: The ID of the agent. (required)
+        :param run_id: The ID of the run. (required)
         :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
@@ -711,7 +710,7 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_serialize(
+        _param = self._get_run_threadstate_serialize(
             run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
@@ -720,16 +719,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "object",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -737,9 +736,9 @@ class RunsApi:
 
 
     @validate_call
-    def get_run_without_preload_content(
+    async def get_run_threadstate_without_preload_content(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
+        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -753,11 +752,11 @@ class RunsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Get a previously created Run
+        """Retrieve the thread state at the end of the run
 
-        Get a run from its ID. Don't wait for the final run output.
+        This call can be used only for agents that support thread, i.e. for Runs that specify a thread ID. It can be called only on runs that are in `success` status. It returns the thread state at the end of the Run. Can be used to reconstruct the evolution of the thread state in its history.
 
-        :param run_id: The ID of the agent. (required)
+        :param run_id: The ID of the run. (required)
         :type run_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
@@ -781,7 +780,7 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_serialize(
+        _param = self._get_run_threadstate_serialize(
             run_id=run_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
@@ -790,19 +789,19 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
+            '200': "object",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
         return response_data.response
 
 
-    def _get_run_serialize(
+    def _get_run_threadstate_serialize(
         self,
         run_id,
         _request_auth,
@@ -849,7 +848,7 @@ class RunsApi:
 
         return self.api_client.param_serialize(
             method='GET',
-            resource_path='/runs/{run_id}',
+            resource_path='/runs/{run_id}/threadstate',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -866,10 +865,9 @@ class RunsApi:
 
 
     @validate_call
-    def get_run_output(
+    async def get_thread(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
-        block_timeout: Optional[StrictInt] = None,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -882,15 +880,13 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> RunOutput:
-        """Retrieve last output of a run if available
+    ) -> Thread:
+        """Get a previously created Thread
 
-        Retrieve the last output of the run.  The output can be:   * an interrupt, this happens when the agent run status is `interrupted`   * the final result of the run, this happens when the agent run status is `success`   * an error, this happens when the agent run status is `error` or `timeout`   If the block timeout is provided and the current run status is `pending`, this call blocks until the state changes or the timeout expires.  If no timeout is provided or the timeout has expired and  run status is `pending`, this call returns `204` with no content.
+        Get a thread from its ID. 
 
-        :param run_id: The ID of the run. (required)
-        :type run_id: str
-        :param block_timeout:
-        :type block_timeout: int
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -913,9 +909,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_output_serialize(
-            run_id=run_id,
-            block_timeout=block_timeout,
+        _param = self._get_thread_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -923,17 +918,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "RunOutput",
-            '204': None,
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -941,10 +935,9 @@ class RunsApi:
 
 
     @validate_call
-    def get_run_output_with_http_info(
+    async def get_thread_with_http_info(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
-        block_timeout: Optional[StrictInt] = None,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -957,15 +950,13 @@ class RunsApi:
         _content_type: Optional[StrictStr] = None,
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[RunOutput]:
-        """Retrieve last output of a run if available
+    ) -> ApiResponse[Thread]:
+        """Get a previously created Thread
 
-        Retrieve the last output of the run.  The output can be:   * an interrupt, this happens when the agent run status is `interrupted`   * the final result of the run, this happens when the agent run status is `success`   * an error, this happens when the agent run status is `error` or `timeout`   If the block timeout is provided and the current run status is `pending`, this call blocks until the state changes or the timeout expires.  If no timeout is provided or the timeout has expired and  run status is `pending`, this call returns `204` with no content.
+        Get a thread from its ID. 
 
-        :param run_id: The ID of the run. (required)
-        :type run_id: str
-        :param block_timeout:
-        :type block_timeout: int
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -988,9 +979,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_output_serialize(
-            run_id=run_id,
-            block_timeout=block_timeout,
+        _param = self._get_thread_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -998,17 +988,16 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "RunOutput",
-            '204': None,
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -1016,10 +1005,9 @@ class RunsApi:
 
 
     @validate_call
-    def get_run_output_without_preload_content(
+    async def get_thread_without_preload_content(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
-        block_timeout: Optional[StrictInt] = None,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1033,14 +1021,12 @@ class RunsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Retrieve last output of a run if available
+        """Get a previously created Thread
 
-        Retrieve the last output of the run.  The output can be:   * an interrupt, this happens when the agent run status is `interrupted`   * the final result of the run, this happens when the agent run status is `success`   * an error, this happens when the agent run status is `error` or `timeout`   If the block timeout is provided and the current run status is `pending`, this call blocks until the state changes or the timeout expires.  If no timeout is provided or the timeout has expired and  run status is `pending`, this call returns `204` with no content.
+        Get a thread from its ID. 
 
-        :param run_id: The ID of the run. (required)
-        :type run_id: str
-        :param block_timeout:
-        :type block_timeout: int
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1063,9 +1049,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._get_run_output_serialize(
-            run_id=run_id,
-            block_timeout=block_timeout,
+        _param = self._get_thread_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1073,23 +1058,21 @@ class RunsApi:
         )
 
         _response_types_map: Dict[str, Optional[str]] = {
-            '200': "RunOutput",
-            '204': None,
+            '200': "Thread",
             '404': "str",
             '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
         return response_data.response
 
 
-    def _get_run_output_serialize(
+    def _get_thread_serialize(
         self,
-        run_id,
-        block_timeout,
+        thread_id,
         _request_auth,
         _content_type,
         _headers,
@@ -1111,13 +1094,9 @@ class RunsApi:
         _body_params: Optional[bytes] = None
 
         # process the path parameters
-        if run_id is not None:
-            _path_params['run_id'] = run_id
+        if thread_id is not None:
+            _path_params['thread_id'] = thread_id
         # process the query parameters
-        if block_timeout is not None:
-            
-            _query_params.append(('block_timeout', block_timeout))
-            
         # process the header parameters
         # process the form parameters
         # process the body parameter
@@ -1138,7 +1117,7 @@ class RunsApi:
 
         return self.api_client.param_serialize(
             method='GET',
-            resource_path='/runs/{run_id}/output',
+            resource_path='/threads/{thread_id}',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
@@ -1155,579 +1134,9 @@ class RunsApi:
 
 
     @validate_call
-    def get_run_stream(
+    async def get_thread_history(
         self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> RunOutputStream:
-        """Stream the run output
-
-        Send a stream of events using Server-sent events (SEE). See <https://html.spec.whatwg.org/multipage/server-sent-events.html> for details.
-
-        :param run_id: The ID of the run. (required)
-        :type run_id: str
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._get_run_stream_serialize(
-            run_id=run_id,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "RunOutputStream",
-            '204': None,
-            '404': "str",
-            '409': "str",
-            '422': "str",
-        }
-        response_data = self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        response_data.read()
-        return self.api_client.response_deserialize(
-            response_data=response_data,
-            response_types_map=_response_types_map,
-        ).data
-
-
-    @validate_call
-    def get_run_stream_with_http_info(
-        self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[RunOutputStream]:
-        """Stream the run output
-
-        Send a stream of events using Server-sent events (SEE). See <https://html.spec.whatwg.org/multipage/server-sent-events.html> for details.
-
-        :param run_id: The ID of the run. (required)
-        :type run_id: str
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._get_run_stream_serialize(
-            run_id=run_id,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "RunOutputStream",
-            '204': None,
-            '404': "str",
-            '409': "str",
-            '422': "str",
-        }
-        response_data = self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        response_data.read()
-        return self.api_client.response_deserialize(
-            response_data=response_data,
-            response_types_map=_response_types_map,
-        )
-
-
-    @validate_call
-    def get_run_stream_without_preload_content(
-        self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the run.")],
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> RESTResponseType:
-        """Stream the run output
-
-        Send a stream of events using Server-sent events (SEE). See <https://html.spec.whatwg.org/multipage/server-sent-events.html> for details.
-
-        :param run_id: The ID of the run. (required)
-        :type run_id: str
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._get_run_stream_serialize(
-            run_id=run_id,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "RunOutputStream",
-            '204': None,
-            '404': "str",
-            '409': "str",
-            '422': "str",
-        }
-        response_data = self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        return response_data.response
-
-
-    def _get_run_stream_serialize(
-        self,
-        run_id,
-        _request_auth,
-        _content_type,
-        _headers,
-        _host_index,
-    ) -> RequestSerialized:
-
-        _host = None
-
-        _collection_formats: Dict[str, str] = {
-        }
-
-        _path_params: Dict[str, str] = {}
-        _query_params: List[Tuple[str, str]] = []
-        _header_params: Dict[str, Optional[str]] = _headers or {}
-        _form_params: List[Tuple[str, str]] = []
-        _files: Dict[
-            str, Union[str, bytes, List[str], List[bytes], List[Tuple[str, bytes]]]
-        ] = {}
-        _body_params: Optional[bytes] = None
-
-        # process the path parameters
-        if run_id is not None:
-            _path_params['run_id'] = run_id
-        # process the query parameters
-        # process the header parameters
-        # process the form parameters
-        # process the body parameter
-
-
-        # set the HTTP header `Accept`
-        if 'Accept' not in _header_params:
-            _header_params['Accept'] = self.api_client.select_header_accept(
-                [
-                    'text/event-stream', 
-                    'application/json'
-                ]
-            )
-
-
-        # authentication setting
-        _auth_settings: List[str] = [
-        ]
-
-        return self.api_client.param_serialize(
-            method='GET',
-            resource_path='/runs/{run_id}/stream',
-            path_params=_path_params,
-            query_params=_query_params,
-            header_params=_header_params,
-            body=_body_params,
-            post_params=_form_params,
-            files=_files,
-            auth_settings=_auth_settings,
-            collection_formats=_collection_formats,
-            _host=_host,
-            _request_auth=_request_auth
-        )
-
-
-
-
-    @validate_call
-    def resume_run(
-        self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
-        body: Dict[str, Any],
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> Run:
-        """Resume an interrupted Run
-
-        Provide the needed input to a run to resume its execution. Can only be called for runs that are in the interrupted state Schema of the provided input must match with the schema specified in the agent specs under interrupts for the interrupt type the agent generated for this specific interruption.
-
-        :param run_id: The ID of the agent. (required)
-        :type run_id: str
-        :param body: (required)
-        :type body: object
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._resume_run_serialize(
-            run_id=run_id,
-            body=body,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
-            '404': "str",
-            '409': "str",
-            '422': "str",
-        }
-        response_data = self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        response_data.read()
-        return self.api_client.response_deserialize(
-            response_data=response_data,
-            response_types_map=_response_types_map,
-        ).data
-
-
-    @validate_call
-    def resume_run_with_http_info(
-        self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
-        body: Dict[str, Any],
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> ApiResponse[Run]:
-        """Resume an interrupted Run
-
-        Provide the needed input to a run to resume its execution. Can only be called for runs that are in the interrupted state Schema of the provided input must match with the schema specified in the agent specs under interrupts for the interrupt type the agent generated for this specific interruption.
-
-        :param run_id: The ID of the agent. (required)
-        :type run_id: str
-        :param body: (required)
-        :type body: object
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._resume_run_serialize(
-            run_id=run_id,
-            body=body,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
-            '404': "str",
-            '409': "str",
-            '422': "str",
-        }
-        response_data = self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        response_data.read()
-        return self.api_client.response_deserialize(
-            response_data=response_data,
-            response_types_map=_response_types_map,
-        )
-
-
-    @validate_call
-    def resume_run_without_preload_content(
-        self,
-        run_id: Annotated[StrictStr, Field(description="The ID of the agent.")],
-        body: Dict[str, Any],
-        _request_timeout: Union[
-            None,
-            Annotated[StrictFloat, Field(gt=0)],
-            Tuple[
-                Annotated[StrictFloat, Field(gt=0)],
-                Annotated[StrictFloat, Field(gt=0)]
-            ]
-        ] = None,
-        _request_auth: Optional[Dict[StrictStr, Any]] = None,
-        _content_type: Optional[StrictStr] = None,
-        _headers: Optional[Dict[StrictStr, Any]] = None,
-        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
-    ) -> RESTResponseType:
-        """Resume an interrupted Run
-
-        Provide the needed input to a run to resume its execution. Can only be called for runs that are in the interrupted state Schema of the provided input must match with the schema specified in the agent specs under interrupts for the interrupt type the agent generated for this specific interruption.
-
-        :param run_id: The ID of the agent. (required)
-        :type run_id: str
-        :param body: (required)
-        :type body: object
-        :param _request_timeout: timeout setting for this request. If one
-                                 number provided, it will be total request
-                                 timeout. It can also be a pair (tuple) of
-                                 (connection, read) timeouts.
-        :type _request_timeout: int, tuple(int, int), optional
-        :param _request_auth: set to override the auth_settings for an a single
-                              request; this effectively ignores the
-                              authentication in the spec for a single request.
-        :type _request_auth: dict, optional
-        :param _content_type: force content-type for the request.
-        :type _content_type: str, Optional
-        :param _headers: set to override the headers for a single
-                         request; this effectively ignores the headers
-                         in the spec for a single request.
-        :type _headers: dict, optional
-        :param _host_index: set to override the host_index for a single
-                            request; this effectively ignores the host_index
-                            in the spec for a single request.
-        :type _host_index: int, optional
-        :return: Returns the result object.
-        """ # noqa: E501
-
-        _param = self._resume_run_serialize(
-            run_id=run_id,
-            body=body,
-            _request_auth=_request_auth,
-            _content_type=_content_type,
-            _headers=_headers,
-            _host_index=_host_index
-        )
-
-        _response_types_map: Dict[str, Optional[str]] = {
-            '200': "Run",
-            '404': "str",
-            '409': "str",
-            '422': "str",
-        }
-        response_data = self.api_client.call_api(
-            *_param,
-            _request_timeout=_request_timeout
-        )
-        return response_data.response
-
-
-    def _resume_run_serialize(
-        self,
-        run_id,
-        body,
-        _request_auth,
-        _content_type,
-        _headers,
-        _host_index,
-    ) -> RequestSerialized:
-
-        _host = None
-
-        _collection_formats: Dict[str, str] = {
-        }
-
-        _path_params: Dict[str, str] = {}
-        _query_params: List[Tuple[str, str]] = []
-        _header_params: Dict[str, Optional[str]] = _headers or {}
-        _form_params: List[Tuple[str, str]] = []
-        _files: Dict[
-            str, Union[str, bytes, List[str], List[bytes], List[Tuple[str, bytes]]]
-        ] = {}
-        _body_params: Optional[bytes] = None
-
-        # process the path parameters
-        if run_id is not None:
-            _path_params['run_id'] = run_id
-        # process the query parameters
-        # process the header parameters
-        # process the form parameters
-        # process the body parameter
-        if body is not None:
-            _body_params = body
-
-
-        # set the HTTP header `Accept`
-        if 'Accept' not in _header_params:
-            _header_params['Accept'] = self.api_client.select_header_accept(
-                [
-                    'application/json'
-                ]
-            )
-
-        # set the HTTP header `Content-Type`
-        if _content_type:
-            _header_params['Content-Type'] = _content_type
-        else:
-            _default_content_type = (
-                self.api_client.select_header_content_type(
-                    [
-                        'application/json'
-                    ]
-                )
-            )
-            if _default_content_type is not None:
-                _header_params['Content-Type'] = _default_content_type
-
-        # authentication setting
-        _auth_settings: List[str] = [
-        ]
-
-        return self.api_client.param_serialize(
-            method='POST',
-            resource_path='/runs/{run_id}',
-            path_params=_path_params,
-            query_params=_query_params,
-            header_params=_header_params,
-            body=_body_params,
-            post_params=_form_params,
-            files=_files,
-            auth_settings=_auth_settings,
-            collection_formats=_collection_formats,
-            _host=_host,
-            _request_auth=_request_auth
-        )
-
-
-
-
-    @validate_call
-    def search_runs(
-        self,
-        run_search_request: RunSearchRequest,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1741,12 +1150,12 @@ class RunsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> List[Run]:
-        """Search Runs
+        """Retrieve the list of runs and associated state at the end of each run.
 
-        Search for runs.  This endpoint also functions as the endpoint to list all runs.
+        Retrieve ordered list of runs for this thread in chronological order.
 
-        :param run_search_request: (required)
-        :type run_search_request: RunSearchRequest
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1769,8 +1178,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._search_runs_serialize(
-            run_search_request=run_search_request,
+        _param = self._get_thread_history_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1779,13 +1188,15 @@ class RunsApi:
 
         _response_types_map: Dict[str, Optional[str]] = {
             '200': "List[Run]",
+            '404': "str",
+            '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -1793,9 +1204,9 @@ class RunsApi:
 
 
     @validate_call
-    def search_runs_with_http_info(
+    async def get_thread_history_with_http_info(
         self,
-        run_search_request: RunSearchRequest,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1809,12 +1220,12 @@ class RunsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> ApiResponse[List[Run]]:
-        """Search Runs
+        """Retrieve the list of runs and associated state at the end of each run.
 
-        Search for runs.  This endpoint also functions as the endpoint to list all runs.
+        Retrieve ordered list of runs for this thread in chronological order.
 
-        :param run_search_request: (required)
-        :type run_search_request: RunSearchRequest
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1837,8 +1248,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._search_runs_serialize(
-            run_search_request=run_search_request,
+        _param = self._get_thread_history_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1847,13 +1258,15 @@ class RunsApi:
 
         _response_types_map: Dict[str, Optional[str]] = {
             '200': "List[Run]",
+            '404': "str",
+            '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
-        response_data.read()
+        await response_data.read()
         return self.api_client.response_deserialize(
             response_data=response_data,
             response_types_map=_response_types_map,
@@ -1861,9 +1274,9 @@ class RunsApi:
 
 
     @validate_call
-    def search_runs_without_preload_content(
+    async def get_thread_history_without_preload_content(
         self,
-        run_search_request: RunSearchRequest,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
         _request_timeout: Union[
             None,
             Annotated[StrictFloat, Field(gt=0)],
@@ -1877,12 +1290,12 @@ class RunsApi:
         _headers: Optional[Dict[StrictStr, Any]] = None,
         _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
     ) -> RESTResponseType:
-        """Search Runs
+        """Retrieve the list of runs and associated state at the end of each run.
 
-        Search for runs.  This endpoint also functions as the endpoint to list all runs.
+        Retrieve ordered list of runs for this thread in chronological order.
 
-        :param run_search_request: (required)
-        :type run_search_request: RunSearchRequest
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
         :param _request_timeout: timeout setting for this request. If one
                                  number provided, it will be total request
                                  timeout. It can also be a pair (tuple) of
@@ -1905,8 +1318,8 @@ class RunsApi:
         :return: Returns the result object.
         """ # noqa: E501
 
-        _param = self._search_runs_serialize(
-            run_search_request=run_search_request,
+        _param = self._get_thread_history_serialize(
+            thread_id=thread_id,
             _request_auth=_request_auth,
             _content_type=_content_type,
             _headers=_headers,
@@ -1915,18 +1328,552 @@ class RunsApi:
 
         _response_types_map: Dict[str, Optional[str]] = {
             '200': "List[Run]",
+            '404': "str",
+            '409': "str",
             '422': "str",
         }
-        response_data = self.api_client.call_api(
+        response_data = await self.api_client.call_api(
             *_param,
             _request_timeout=_request_timeout
         )
         return response_data.response
 
 
-    def _search_runs_serialize(
+    def _get_thread_history_serialize(
         self,
-        run_search_request,
+        thread_id,
+        _request_auth,
+        _content_type,
+        _headers,
+        _host_index,
+    ) -> RequestSerialized:
+
+        _host = None
+
+        _collection_formats: Dict[str, str] = {
+        }
+
+        _path_params: Dict[str, str] = {}
+        _query_params: List[Tuple[str, str]] = []
+        _header_params: Dict[str, Optional[str]] = _headers or {}
+        _form_params: List[Tuple[str, str]] = []
+        _files: Dict[
+            str, Union[str, bytes, List[str], List[bytes], List[Tuple[str, bytes]]]
+        ] = {}
+        _body_params: Optional[bytes] = None
+
+        # process the path parameters
+        if thread_id is not None:
+            _path_params['thread_id'] = thread_id
+        # process the query parameters
+        # process the header parameters
+        # process the form parameters
+        # process the body parameter
+
+
+        # set the HTTP header `Accept`
+        if 'Accept' not in _header_params:
+            _header_params['Accept'] = self.api_client.select_header_accept(
+                [
+                    'application/json'
+                ]
+            )
+
+
+        # authentication setting
+        _auth_settings: List[str] = [
+        ]
+
+        return self.api_client.param_serialize(
+            method='GET',
+            resource_path='/threads/{thread_id}/history',
+            path_params=_path_params,
+            query_params=_query_params,
+            header_params=_header_params,
+            body=_body_params,
+            post_params=_form_params,
+            files=_files,
+            auth_settings=_auth_settings,
+            collection_formats=_collection_formats,
+            _host=_host,
+            _request_auth=_request_auth
+        )
+
+
+
+
+    @validate_call
+    async def get_thread_state(
+        self,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> object:
+        """Retrieve the current state associated with the thread
+
+        Retrieve the the current state associated with the thread
+
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._get_thread_state_serialize(
+            thread_id=thread_id,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "object",
+            '404': "str",
+            '409': "str",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        await response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        ).data
+
+
+    @validate_call
+    async def get_thread_state_with_http_info(
+        self,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> ApiResponse[object]:
+        """Retrieve the current state associated with the thread
+
+        Retrieve the the current state associated with the thread
+
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._get_thread_state_serialize(
+            thread_id=thread_id,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "object",
+            '404': "str",
+            '409': "str",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        await response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        )
+
+
+    @validate_call
+    async def get_thread_state_without_preload_content(
+        self,
+        thread_id: Annotated[StrictStr, Field(description="The ID of the thread.")],
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> RESTResponseType:
+        """Retrieve the current state associated with the thread
+
+        Retrieve the the current state associated with the thread
+
+        :param thread_id: The ID of the thread. (required)
+        :type thread_id: str
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._get_thread_state_serialize(
+            thread_id=thread_id,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "object",
+            '404': "str",
+            '409': "str",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        return response_data.response
+
+
+    def _get_thread_state_serialize(
+        self,
+        thread_id,
+        _request_auth,
+        _content_type,
+        _headers,
+        _host_index,
+    ) -> RequestSerialized:
+
+        _host = None
+
+        _collection_formats: Dict[str, str] = {
+        }
+
+        _path_params: Dict[str, str] = {}
+        _query_params: List[Tuple[str, str]] = []
+        _header_params: Dict[str, Optional[str]] = _headers or {}
+        _form_params: List[Tuple[str, str]] = []
+        _files: Dict[
+            str, Union[str, bytes, List[str], List[bytes], List[Tuple[str, bytes]]]
+        ] = {}
+        _body_params: Optional[bytes] = None
+
+        # process the path parameters
+        if thread_id is not None:
+            _path_params['thread_id'] = thread_id
+        # process the query parameters
+        # process the header parameters
+        # process the form parameters
+        # process the body parameter
+
+
+        # set the HTTP header `Accept`
+        if 'Accept' not in _header_params:
+            _header_params['Accept'] = self.api_client.select_header_accept(
+                [
+                    'application/json'
+                ]
+            )
+
+
+        # authentication setting
+        _auth_settings: List[str] = [
+        ]
+
+        return self.api_client.param_serialize(
+            method='GET',
+            resource_path='/threads/{thread_id}/state',
+            path_params=_path_params,
+            query_params=_query_params,
+            header_params=_header_params,
+            body=_body_params,
+            post_params=_form_params,
+            files=_files,
+            auth_settings=_auth_settings,
+            collection_formats=_collection_formats,
+            _host=_host,
+            _request_auth=_request_auth
+        )
+
+
+
+
+    @validate_call
+    async def search_threads(
+        self,
+        thread_search_request: ThreadSearchRequest,
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> List[Thread]:
+        """Search Threads
+
+        Search for threads.  This endpoint also functions as the endpoint to list all threads.
+
+        :param thread_search_request: (required)
+        :type thread_search_request: ThreadSearchRequest
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._search_threads_serialize(
+            thread_search_request=thread_search_request,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "List[Thread]",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        await response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        ).data
+
+
+    @validate_call
+    async def search_threads_with_http_info(
+        self,
+        thread_search_request: ThreadSearchRequest,
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> ApiResponse[List[Thread]]:
+        """Search Threads
+
+        Search for threads.  This endpoint also functions as the endpoint to list all threads.
+
+        :param thread_search_request: (required)
+        :type thread_search_request: ThreadSearchRequest
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._search_threads_serialize(
+            thread_search_request=thread_search_request,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "List[Thread]",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        await response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        )
+
+
+    @validate_call
+    async def search_threads_without_preload_content(
+        self,
+        thread_search_request: ThreadSearchRequest,
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[
+                Annotated[StrictFloat, Field(gt=0)],
+                Annotated[StrictFloat, Field(gt=0)]
+            ]
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> RESTResponseType:
+        """Search Threads
+
+        Search for threads.  This endpoint also functions as the endpoint to list all threads.
+
+        :param thread_search_request: (required)
+        :type thread_search_request: ThreadSearchRequest
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """ # noqa: E501
+
+        _param = self._search_threads_serialize(
+            thread_search_request=thread_search_request,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            '200': "List[Thread]",
+            '422': "str",
+        }
+        response_data = await self.api_client.call_api(
+            *_param,
+            _request_timeout=_request_timeout
+        )
+        return response_data.response
+
+
+    def _search_threads_serialize(
+        self,
+        thread_search_request,
         _request_auth,
         _content_type,
         _headers,
@@ -1952,8 +1899,8 @@ class RunsApi:
         # process the header parameters
         # process the form parameters
         # process the body parameter
-        if run_search_request is not None:
-            _body_params = run_search_request
+        if thread_search_request is not None:
+            _body_params = thread_search_request
 
 
         # set the HTTP header `Accept`
@@ -1984,7 +1931,7 @@ class RunsApi:
 
         return self.api_client.param_serialize(
             method='POST',
-            resource_path='/runs/search',
+            resource_path='/threads/search',
             path_params=_path_params,
             query_params=_query_params,
             header_params=_header_params,
